requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
PyGithub>=2.1.1
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.3
tiktoken>=0.5.2
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
rich>=13.7.0
click>=8.1.7
PyYAML>=6.0.1
